Move scene data structure functions into hooks file
