+++ src/panels/Scene/SceneTree/SceneTree.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useState } from 'react';
 import {
   ActionIcon,
   Box,
@@ -14,19 +14,14 @@ import { generateMatcherFunctionByKeys } from '@/components/FilterList/util';
 import { ChevronsDownIcon, ChevronsUpIcon } from '@/icons/icons';
 import { useAppDispatch, useAppSelector } from '@/redux/hooks';
 import { storeSceneTreeNodeExpanded } from '@/redux/local/localSlice';
-import { SceneTreeGroupPrefixKey } from '@/util/sceneTreeGroupsHelper';
+import { SceneTreeGroupPrefixKey } from '@/util/groupsHelpers';
 
 import { useOpenCurrentSceneNodeWindow } from '../hooks';
 
+import { useSceneTreeData } from './hooks';
 import { SceneTreeFilters } from './SceneTreeFilters';
 import { SceneTreeNode, SceneTreeNodeContent } from './SceneTreeNode';
-import {
-  filterTreeData,
-  flattenTreeData,
-  SceneTreeFilterSettings,
-  sortTreeData
-} from './treeUtil';
-import { SceneTreeNodeData } from './types';
+import { SceneTreeFilterSettings, SceneTreeNodeData } from './types';
 
 export function SceneTree() {
   const [filter, setFilter] = useState<SceneTreeFilterSettings>({
@@ -35,23 +30,12 @@ export function SceneTree() {
     tags: []
   });
 
+  const { sceneTreeData, flatTreeData } = useSceneTreeData(filter);
   const { closeCurrentNodeWindow } = useOpenCurrentSceneNodeWindow();
-
-  const sceneTreeData = useAppSelector((state) => state.groups.sceneTreeData);
-  // @TODO (emmbr, 2024-12-17): Remove dependency on entire properties object. This means
-  // that the entire menu is rerendered as soon as a property changes... Alternatively,
-  // update state structure so that the property values are stored in a separate object
-  const properties = useAppSelector((state) => state.properties.properties);
-  const propertyOwners = useAppSelector((state) => state.propertyOwners.propertyOwners);
-  const customGuiOrdering = useAppSelector((state) => state.groups.customGroupOrdering);
 
   const initialExpandedNodes = useAppSelector(
     (state) => state.local.sceneTree.expandedGroups
   );
-
-  function isGroup(value: string): boolean {
-    return value.startsWith(SceneTreeGroupPrefixKey);
-  }
 
   const tree = useTree({
     initialExpandedState: getTreeExpandedState(sceneTreeData, initialExpandedNodes)
@@ -59,40 +43,25 @@ export function SceneTree() {
 
   const dispatch = useAppDispatch();
 
-  // This will only be run on unmount
+  // This will run after each render to update the expanded state in redux
   useEffect(() => {
     return () => {
       // Convert the map to a list of strings with keys of the
       // expanded groups
       const expandedGroups = Object.entries(tree.expandedState)
-        .filter(([key, isOpen]) => isGroup(key) && isOpen)
+        .filter(([key, isOpen]) => key.startsWith(SceneTreeGroupPrefixKey) && isOpen)
         .map(([key]) => key); // Save expanded state on unmount
 
       dispatch(storeSceneTreeNodeExpanded(expandedGroups));
-      // Also close the "current node" window
+    };
+  }, [dispatch, tree.expandedState]);
+
+  // This will only be run on unmount to close the window of the opened node
+  useEffect(() => {
+    return () => {
       closeCurrentNodeWindow();
     };
-  }, [dispatch, closeCurrentNodeWindow, tree.expandedState]);
-
-  // If the sorting is not wrapped in an useMemo, the component will rerender constantly,
-  // for some reason
-  const treeData = useMemo(() => {
-    let data = sceneTreeData;
-    data = filterTreeData(data, filter, properties, propertyOwners);
-    data = sortTreeData(data, customGuiOrdering, properties);
-    return data;
-  }, [customGuiOrdering, filter, properties, propertyOwners, sceneTreeData]);
-
-  // Create a flat list of all leaf nodes, that we can use for searching
-  const flatTreeData = flattenTreeData(treeData);
-
-  // @TODO (2025-01-13 emmbr): Would be nice to sort the results by some type of
-  // "relevance", but for now we just sort alphabetically
-  flatTreeData.sort((a: SceneTreeNodeData, b: SceneTreeNodeData) => {
-    const nameA = a.label as string;
-    const nameB = b.label as string;
-    return nameA.toLocaleLowerCase().localeCompare(nameB.toLocaleLowerCase());
-  });
+  }, [closeCurrentNodeWindow]);
 
   return (
     <FilterList>
@@ -119,7 +88,7 @@ export function SceneTree() {
           </Group>
         </Box>
         <Tree
-          data={treeData}
+          data={sceneTreeData}
           tree={tree}
           renderNode={(payload) => <SceneTreeNode {...payload} />}
         />
