+++ src/redux/groups/groupsSlice.ts
@@ -1,19 +1,16 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
-import { SceneTreeNodeData } from '@/panels/Scene/SceneTree/types';
 import { CustomGroupOrdering, Groups } from '@/types/types';
 
 export interface GroupsState {
   customGroupOrdering: CustomGroupOrdering;
   groups: Groups;
-  sceneTreeData: SceneTreeNodeData[];
   tags: string[];
 }
 
 const initialState: GroupsState = {
   customGroupOrdering: {},
   groups: {},
-  sceneTreeData: [],
   tags: []
 };
 
@@ -23,10 +20,6 @@ export const groupsSlice = createSlice({
   reducers: {
     setGroups: (state, action: PayloadAction<Groups>) => {
       state.groups = action.payload;
-      return state;
-    },
-    setSceneTreeData: (state, action: PayloadAction<SceneTreeNodeData[]>) => {
-      state.sceneTreeData = action.payload;
       return state;
     },
     setTags: (state, action: PayloadAction<string[]>) => {
@@ -41,6 +34,5 @@ export const groupsSlice = createSlice({
 });
 
 // Action creators are generated for each case reducer function, replaces the `Actions/index.js`
-export const { setGroups, setSceneTreeData, setTags, updateCustomGroupOrdering } =
-  groupsSlice.actions;
+export const { setGroups, setTags, updateCustomGroupOrdering } = groupsSlice.actions;
 export const groupsReducer = groupsSlice.reducer;
