+++ src/util/groupsHelpers.ts
@@ -62,83 +62,3 @@ export function computeGroups(
 
   return groups;
 }
-
-/**
- * Create the data for a Scene tree node representing a scene graph node.
- */
-export function treeDataForSceneGraphNode(
-  name: string,
-  uri: string,
-  path?: string
-): SceneTreeNodeData {
-  return {
-    label: name || '',
-    value: uri,
-    guiPath: path?.split('/') ?? []
-  };
-}
-
-/**
- * Create the data for the Scene tree from the groups information.
- */
-export function sceneTreeDataFromGroups(
-  groups: Groups,
-  propertyOwners: PropertyOwners
-): SceneTreeNodeData[] {
-  const treeData: SceneTreeNodeData[] = [];
-
-  const topLevelGroupsPaths = Object.keys(groups).filter((path) => {
-    // Get the number of slashes in the path
-    const depth = (path.match(/\//g) || []).length;
-    return depth === 1 && path !== '/';
-  });
-
-  // Build the data structure for the tree
-  function generateGroupData(path: string) {
-    const splitPath = path.split('/');
-    const name = splitPath.length > 1 ? splitPath.pop() : 'Untitled';
-
-    const groupNodeData: SceneTreeNodeData = {
-      value: SceneTreeGroupPrefixKey + path,
-      label: name,
-      children: [],
-      guiPath: path.split('/')
-    };
-
-    const groupData = groups[path];
-
-    // Add subgroups, recursively
-    groupData.subgroups.forEach((subGroupPath) =>
-      groupNodeData.children?.push(generateGroupData(subGroupPath))
-    );
-
-    // Add property owners, also recursively
-    groupData.propertyOwners.forEach((uri) => {
-      const owner = propertyOwners[uri];
-      if (owner === undefined) {
-        return;
-      }
-      groupNodeData.children?.push(
-        treeDataForSceneGraphNode(owner.name, owner.uri, path)
-      );
-    });
-
-    return groupNodeData;
-  }
-
-  topLevelGroupsPaths.forEach((path) => {
-    treeData.push(generateGroupData(path));
-  });
-
-  // Add the nodes without any group to the top level
-  const nodesWithoutGroup = groups['/']?.propertyOwners || [];
-  nodesWithoutGroup.forEach((uri) => {
-    const owner = propertyOwners[uri];
-    if (owner === undefined) {
-      return;
-    }
-    treeData.push(treeDataForSceneGraphNode(owner.name, owner.uri));
-  });
-
-  return treeData;
-}
